MOVE: moved slatekit.common.functions to new slatekit.functions project
