+++ src/lib/kotlin/slatekit-functions/src/main/kotlin/slatekit/functions/cmds/CommandResult.kt
@@ -14,9 +14,7 @@
 package slatekit.functions.cmds
 
 import slatekit.common.DateTime
-import slatekit.common.functions.FunctionInfo
-import slatekit.common.functions.FunctionMode
-import slatekit.common.functions.FunctionResult
+import slatekit.functions.common.*
 import slatekit.results.Result
 import slatekit.results.builders.Tries
