+++ src/lib/kotlin/slatekit-functions/src/main/kotlin/slatekit/functions/cmds/CommandState.kt
@@ -16,9 +16,7 @@ package slatekit.functions.cmds
 import slatekit.common.DateTime
 import slatekit.common.DateTimes
 import slatekit.common.Status
-import slatekit.common.functions.FunctionInfo
-import slatekit.common.functions.FunctionMode
-import slatekit.common.functions.FunctionState
+import slatekit.functions.common.*
 import slatekit.common.metrics.Metrics
 import slatekit.common.metrics.MetricsLite
