+++ src/lib/kotlin/slatekit-functions/src/main/kotlin/slatekit/functions/common/FunctionInfo.kt
@@ -1,4 +1,4 @@
-package slatekit.common.functions
+package slatekit.functions.common
 
 import slatekit.common.Random
 import slatekit.common.toId
