+++ src/lib/kotlin/slatekit-functions/src/main/kotlin/slatekit/functions/common/FunctionTriggers.kt
@@ -1,4 +1,4 @@
-package slatekit.common.functions
+package slatekit.functions.common
 
 
 interface FunctionTriggers<out Result> {
