+++ src/lib/kotlin/slatekit-tests/src/test/kotlin/test/core/SyncTests.kt
@@ -2,8 +2,8 @@ package test.core
 
 import org.junit.Assert
 import org.junit.Test
-import slatekit.common.functions.FunctionMode
-import slatekit.core.syncs.Sync
+import slatekit.functions.common.FunctionMode
+import slatekit.functions.syncs.Sync
 import slatekit.results.builders.Notices
 
 class SyncTests {
